requirements.txt
pillow
google-auth-oauthlib